requirements.txt
appdirs==1.4.3
cvxopt==1.1.9
cycler==0.10.0
fuel==0.2.0
functools32==3.2.3.post2
h5py==2.7.0
Lasagne==0.2.dev1
lmdb==0.93
matplotlib==2.0.0
numexpr==2.6.2
numpy==1.12.1
olefile==0.44
opencv-python==3.3.0.9
packaging==16.8
pandas==0.19.2
picklable-itertools==0.1.1
Pillow==4.1.0
progressbar2==3.18.1
pycrayon==0.5
pyparsing==2.2.0
python-dateutil==2.6.0
python-utils==2.1.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
six==1.10.0
subprocess32==3.2.7
tables==3.4.2
Theano==0.8.2
tqdm==4.11.2
